Template OrderedMultiQueue
